+++ src/main/java/vendingmachine/Application.java
@@ -1,11 +1,17 @@
 package vendingmachine;
 
 import vendingmachine.controller.VendingMachineController;
+import vendingmachine.model.VendingMachine;
 
 public class Application {
 	public static void main(String[] args) {
-		VendingMachineController vendingMachineController = new VendingMachineController();
-		vendingMachineController.use();
-		// TODO: 프로그램 구현
+		VendingMachine vendingMachine = new VendingMachine();
+		VendingMachineController vendingMachineController = new VendingMachineController(vendingMachine);
+
+		vendingMachineController.setInitCoins();
+		vendingMachineController.setMenuList();
+		vendingMachineController.insertMoney();
+		vendingMachineController.orderUntilEnd();
+		vendingMachineController.giveChangesAtTheEnd();
 	}
 }
